webpubsub/chatApp/requirements.txt
azure.messaging.webpubsubservice
dotenv
websockets
flask